scripts/N-star.py
import tensorflow as tf
from tensorflow.keras.models import Model
from tensorflow.keras.layers import Input, Dense, Flatten, Conv2D
from transformers import TFGPT2Model, TFBertModel

import tensorflow as tf
from tensorflow.keras.models import Model
from tensorflow.keras.layers import Input, Dense, Flatten, Conv2D
from transformers import TransformerBlock  # Assume an implemented transformer module



def create_initial_population(module, population_size, parameter_range):
    population = []
    for _ in range(population_size):
        individual = {param: np.random.uniform(low, high) 
                      for param, (low, high) in parameter_range.items()}
        population.append(individual)
    return population


def evaluate_population(module, population, evaluation_metric):
    fitness_scores = []
    for individual in population:
        module.configure(individual)  # Configure the module with the individual's parameters
        score = evaluation_metric(module)  # Evaluate the module
        fitness_scores.append(score)
    return fitness_scores

#region GAN model
class GAN(Model):
    def __init__(self, input_shape, latent_dim):
        super(GAN, self).__init__()
        self.latent_dim = latent_dim
        self.generator = self.create_generator()
        self.discriminator = self.create_discriminator(input_shape)

    def compile(self, gen_optimizer, disc_optimizer, loss_function):
        super(GAN, self).compile()
        self.gen_optimizer = gen_optimizer
        self.disc_optimizer = disc_optimizer
        self.loss_function = loss_function

    @staticmethod
    def create_generator(latent_dim=100):
        # Define generator model architecture
        model = tf.keras.Sequential([
            Dense(256, activation='relu', input_dim=latent_dim),
            Dense(512, activation='relu'),
            Dense(1024, activation='relu'),
            Dense(np.prod(input_shape), activation='tanh'),
            Reshape(input_shape)
        ])
        return model

    @staticmethod
    def create_discriminator(input_shape):
        # Define discriminator model architecture
        model = tf.keras.Sequential([
            Flatten(input_shape=input_shape),
            Dense(512, activation='relu'),
            Dense(256, activation='relu'),
            Dense(1, activation='sigmoid')
        ])
        return model

    def train_step(self, real_images):
        # Training logic for one step
        # Generate fake images
        random_latent_vectors = tf.random.normal(shape=(batch_size, self.latent_dim))
        generated_images = self.generator(random_latent_vectors)

        # Combine real and fake images
        combined_images = tf.concat([generated_images, real_images], axis=0)

        # Labels for fake and real images
        labels = tf.concat([tf.ones((batch_size, 1)), tf.zeros((batch_size, 1))], axis=0)
        labels += 0.05 * tf.random.uniform(tf.shape(labels))  # Add label noise

        # Train the discriminator
        with tf.GradientTape() as tape:
            predictions = self.discriminator(combined_images)
            d_loss = self.loss_function(labels, predictions)
        grads = tape.gradient(d_loss, self.discriminator.trainable_weights)
        self.disc_optimizer.apply_gradients(zip(grads, self.discriminator.trainable_weights))

        # Train the generator
        misleading_labels = tf.zeros((batch_size, 1))
        with tf.GradientTape() as tape:
            predictions = self.discriminator(self.generator(random_latent_vectors))
            g_loss = self.loss_function(misleading_labels, predictions)
        grads = tape.gradient(g_loss, self.generator.trainable_weights)
        self.gen_optimizer.apply_gradients(zip(grads, self.generator.trainable_weights))

        return {'d_loss': d_loss, 'g_loss': g_loss}

    def save_weights(self, filepath):
        self.generator.save_weights(filepath + '_generator.h5')
        self.discriminator.save_weights(filepath + '_discriminator.h5')

    def load_weights(self, filepath):
        self.generator.load_weights(filepath + '_generator.h5')
        self.discriminator.load_weights(filepath + '_discriminator.h5')

#endregion


class EnhancedCNN(Model):
    def __init__(self, input_shape, num_classes):
        super(EnhancedCNN, self).__init__()
        self.model = self.create_model(input_shape, num_classes)

    @staticmethod
    def create_model(input_shape, num_classes):
        model = tf.keras.Sequential([
            Conv2D(filters=64, kernel_size=(3, 3), activation='relu', padding='same', input_shape=input_shape),
            Conv2D(filters=64, kernel_size=(3, 3), activation='relu', padding='same'),
            MaxPooling2D(pool_size=(2, 2)),
            Dropout(0.25),
            Conv2D(filters=128, kernel_size=(3, 3), activation='relu', padding='same'),
            Conv2D(filters=128, kernel_size=(3, 3), activation='relu', padding='same'),
            MaxPooling2D(pool_size=(2, 2)),
            Dropout(0.25),
            Flatten(),
            Dense(512, activation='relu'),
            Dropout(0.5),
            Dense(num_classes, activation='softmax')
        ])
        model.compile(optimizer='adam', loss='categorical_crossentropy', metrics=['accuracy'])
        return model

    def call(self, inputs):
        return self.model(inputs)
    
    def save_model(self, filepath):
        self.model.save(filepath)

    def load_model(self, filepath):
        self.model = tf.keras.models.load_model(filepath)



from transformers import TFGPT2Model, TFBertModel

class AdvancedCEN(Model):
    def __init__(self, model_type='GPT2', input_shape=(224, 224, 3), num_classes=10):
        super(AdvancedCEN, self).__init__()
        if model_type == 'GPT2':
            self.transformer = TFGPT2Model.from_pretrained('gpt2-medium')
        elif model_type == 'BERT':
            self.transformer = TFBertModel.from_pretrained('bert-base-uncased')
        self.flatten = Flatten()
        self.dense1 = Dense(512, activation='relu')
        self.dropout = Dropout(0.5)
        self.output_layer = Dense(num_classes, activation='softmax')

    def call(self, inputs):
        x = self.transformer(inputs)[0]
        x = self.flatten(x)
        x = self.dense1(x)
        x = self.dropout(x)
        return self.output_layer(x)

    def compile(self, optimizer='adam', loss='categorical_crossentropy', metrics=['accuracy']):
        super(AdvancedCEN, self).compile(optimizer=optimizer, loss=loss, metrics=metrics)

    def save_weights(self, filepath):
        self.save_weights(filepath + '.h5')

    def load_weights(self, filepath):
        self.load_weights(filepath + '.h5')

# Update in AGISystem
agi_system.set_cen(AdvancedCEN(model_type='GPT2'))


class DynamicAttentionLayer(tf.keras.layers.Layer):
    def __init__(self, units, **kwargs):
        super(DynamicAttentionLayer, self).__init__(**kwargs)
        self.units = units
        self.context_vector = self.add_weight(name='context_vector',
                                              shape=(self.units, 1),
                                              initializer='uniform',
                                              trainable=True)

    def call(self, inputs):
        inputs_expanded = K.expand_dims(inputs, axis=-1)
        attention_weights = K.softmax(K.dot(inputs, self.context_vector), axis=1)
        attention_weights_expanded = K.expand_dims(attention_weights, axis=-1)
        weighted_output = inputs_expanded * attention_weights_expanded
        weighted_output = K.sum(weighted_output, axis=1)
        return weighted_output


# Spiking Neural Network (SNN) Pathway
class SNNPathway:
    def __init__(self):
        # Define SNN initialization logic
        self.weights = None  # Placeholder for synaptic weights

    def transmit(self, data):
        # SNN data transmission simulation
        processed_data = data * self.weights  # Simplified example
        return processed_data

    def update_weights(self, learning_rate, reward_signal, environmental_data, decay_factor=0.99, plasticity_rule='STDP'):
        """
        Advanced weight update mechanism for synaptic learning, integrating environmental feedback and plasticity rules.

        :param learning_rate: Learning rate for synaptic adjustments.
        :param reward_signal: Reward or penalty signal from the AGI system.
        :param environmental_data: Data from the environment or other modules for contextual learning.
        :param decay_factor: Decay rate for synaptic strength over time.
        :param plasticity_rule: Type of synaptic plasticity rule to apply (e.g., 'STDP' for Spike-Timing-Dependent Plasticity).
        """
        # Synaptic Plasticity based on Spike-Timing-Dependent Plasticity (STDP) or other rules
        if plasticity_rule == 'STDP':
            for i in range(self.num_neurons):
                for j in range(self.num_neurons):
                    if self.last_spike_time[i] < self.last_spike_time[j]:
                        self.weights[i, j] += learning_rate * reward_signal
                    else:
                        self.weights[i, j] -= learning_rate * reward_signal

        # Synaptic Scaling based on environmental feedback
        for i in range(self.num_neurons):
            environmental_factor = self._process_environmental_data(environmental_data[i])
            self.weights[:, i] *= environmental_factor

        # Synaptic Decay over time to simulate natural forgetting and resource optimization
        self.weights *= decay_factor


    def _process_environmental_data(self, data):
        """
        Process environmental data to extract relevant features for synaptic scaling.
        
        :param data: Environmental input data.
        :return: Scaling factor based on environmental data.
        """
        # This is a placeholder for a complex data processing mechanism
        # It could involve feature extraction, filtering, or transformation based on the AGI's current objectives
        return np.mean(data)  # Example: mean value as a simple feature
        # Normalize weights to prevent unbounded growth
        self.weights = np.clip(self.weights, -1, 1)

    def save_weights(self, filepath):
        np.save(filepath + '_snn_weights.npy', self.weights)

    def load_weights(self, filepath):
        self.weights = np.load(filepath + '_snn_weights.npy')


# Placeholder for Spiking Neural Network (SNN) and Q-learning agent
class DynamicQLearningAgent:
    def __init__(self):
        self.model = self.create_q_network_with_attention(state_size, action_size)
        self.gamma = 0.95    # discount rate
        self.epsilon = 1.0   # exploration rate
        self.epsilon_min = 0.01
        self.epsilon_decay = 0.995

    def create_q_network_with_attention(self, state_size, action_size):
        input_layer = Input(shape=(state_size,))
        x = Dense(24, activation='relu')(input_layer)
        x = DynamicAttentionLayer(24)(x)
        x = Dense(24, activation='relu')(x)
        output_layer = Dense(action_size, activation='linear')(x)
        model = Model(inputs=input_layer, outputs=output_layer)
        model.compile(loss='mse', optimizer=Adam(learning_rate=0.001))
        return model

    def save_weights(self, filepath):
        self.model.save_weights(filepath + '_qnetwork.h5')

    def load_weights(self, filepath):
        self.model.load_weights(filepath + '_qnetwork.h5')

# MetaLearningController for adaptive learning
class MetaLearningController:
    def __init__(self, cen, initial_learning_rate=0.01):
        self.cen = cen  # Central Executive Network
        self.learning_rate = initial_learning_rate

    def adapt(self, feedback):
        """
        Dynamically adjusts the learning parameters of the CEN based on received feedback.

        :param feedback: A dictionary containing performance metrics and other feedback indicators.
        """
        performance = feedback.get('performance', 0)
        accuracy = feedback.get('accuracy', 0)

        # Adjust learning rate based on performance and accuracy
        if performance > 0.8 and accuracy > 0.8:
            self.learning_rate *= 0.9  # Decrease learning rate to fine-tune the model
        elif performance < 0.5 or accuracy < 0.5:
            self.learning_rate *= 1.1  # Increase learning rate to accelerate learning

        # Update CEN's learning parameters
        self.cen.optimizer.learning_rate = self.learning_rate



class AdvancedCEN(Model):
    def __init__(self, model_type='GPT2', input_shape=(224, 224, 3)):
        super(AdvancedCEN, self).__init__()
        if model_type == 'GPT2':
            self.transformer = TFGPT2Model.from_pretrained('gpt2')
        elif model_type == 'BERT':
            self.transformer = TFBertModel.from_pretrained('bert-base-uncased')
        self.flatten = Flatten()
        self.output_layer = Dense(units=10, activation='softmax')

    def call(self, inputs):
        x = self.transformer(inputs)[0]
        x = self.flatten(x)
        return self.output_layer(x)


# AGI System with Integrated Components
class AGISystem:
    def __init__(self):
        self.submodules = {}
        self.cen = AdvancedCEN()
        self.snn_q_learning_agent = DynamicQLearningAgent()
        self.snn_pathway = SNNPathway()
        self.snn_q_learning_agent = DynamicQLearningAgent()
        self.meta_learning_controller = MetaLearningController(self.cen)

    def register_submodule(self, name, submodule):
        """ Register a submodule with its capabilities. """
        self.submodules[name] = submodule

    def remove_submodule(self, name):
        """ Remove a submodule from the registry. """
        if name in self.submodules:
            del self.submodules[name]

    def set_cen(self, cen):
        """ Set the Central Executive Network (CEN). """
        self.cen = cen

    def get_submodule(self, name):
        """ Retrieve a submodule by name. """
        return self.submodules.get(name, None)

    # Placeholder for dynamic interactions between submodules and the CEN
    def interact(self):
        pass  # Implement interaction logic

# Instantiate the AGI system
agi_system = AGISystem()


def customize_gan(gan, generator_params, discriminator_params):
    gan.generator = tf.keras.Sequential(generator_params)
    gan.discriminator = tf.keras.Sequential(discriminator_params)
    
    
def evolutionary_optimize(module, optimization_objective):
    population = create_initial_population(module)
    for generation in range(num_generations):
        evaluate_population(population, optimization_objective)
        population = select_and_reproduce(population)
        mutate_population(population)
    best_solution = select_best_individual(population)
    update_module_with_best_solution(module, best_solution)

# Example usage
evolutionary_optimize(cnn, optimization_objective="accuracy")




def main():
    # AGI System setup
    input_shape = (224, 224, 3)
    gan = GAN(input_shape, latent_dim=100)
    cnn = EnhancedCNN(input_shape, num_classes=10)
    cen = AdvancedCEN(model_type='GPT2', input_shape=input_shape)
    agi_system = AGISystem()
    agi_system.set_cen(cen)
    agi_system.register_submodule('GAN', gan)
    agi_system.register_submodule('CNN', cnn)

    load_existing_weights = True # Set based on requirement

    if load_existing_weights:
        try:
            gan.load_weights('gan_weights')
            cnn.load_model('cnn_model')
            cen.load_weights('cen_weights')
            agi_system.snn_pathway.load_weights('snn_pathway_weights')
            print("Existing weights loaded.")
        except Exception as e:
            print(f"Error loading weights: {e}")

    # Main AGI System Loop
    for epoch in range(number_of_epochs):
        # Interact with environment and collect data
        environment_data = collect_environment_data() # Placeholder for data collection logic

        # Process data through various components
        processed_data_cnn = cnn(environment_data)
        processed_data_cen = cen(processed_data_cnn)
        feedback = analyze_results(processed_data_cen) # Placeholder for result analysis logic

        # Update MetaLearningController with feedback
        agi_system.meta_learning_controller.adapt(feedback)

        # Transmit data through SNNPathway and update weights
        snn_output = agi_system.snn_pathway.transmit(environment_data)
        agi_system.snn_pathway.update_weights(learning_rate, reward_signal, snn_output)

        # Train and adapt GAN
        gan.train_step(environment_data)

    # Save weights after training
    gan.save_weights('gan_weights')
    cnn.save_model('cnn_model')
    cen.save_weights('cen_weights')
    agi_system.snn_pathway.save_weights('snn_pathway_weights')
    print("Training complete. Model weights saved.")

if __name__ == "__main__":
    main()